style: 타입, 스타일 변경

+++ src/main/java/com/example/Sparta_Store/orders/dto/response/OrderResponseDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 public record OrderResponseDto(
     String orderId,
-    int totalPrice,
+    long totalPrice,
     OrderStatus orderStatus,
     LocalDateTime createdAt
 ) {
